CLJS-2622: Add support for package tarballs in :npm-deps
This change allows local NPM package tarballs to be imported via
:npm-deps directly, rather than only being able to import published
packages. The syntax is:

  {:npm-deps {"/path/to/package.tgz" "..."}}

This is implemented with a simple conditional: if the dependency
name is a string and points to an existing file, we assume we can
install it with npm install.

Along with the functionality, this commit adds two tests: one to
verify that installing and indexing a tarball package works and one
to verify that building with a tarball package dependency works
as well. Both tests are based on a minimal example JS package, located
in src/tests/cljs_build/npm_deps_tarball_test/example-package-sources.

For these tests to be possible, a small cljs.test-util/run-command!
utility was extracted from cljs.closure/maybe-install-node-deps!.